+++ tla-bmcmt/src/main/scala/at/forsyte/apalache/tla/bmcmt/Z3SolverContext.scala
@@ -2,7 +2,8 @@ package at.forsyte.apalache.tla.bmcmt
 
 import java.io.{File, PrintWriter}
 
-import at.forsyte.apalache.tla.bmcmt.profiler.{FruitlessSmtListener, SmtListener}
+import at.forsyte.apalache.tla.bmcmt.profiler.{IdleSmtListener, SmtListener}
+import at.forsyte.apalache.tla.bmcmt.rewriter.ConstSimplifierForSmt
 import at.forsyte.apalache.tla.bmcmt.types.{BoolT, CellT, FailPredT, IntT}
 import at.forsyte.apalache.tla.lir._
 import at.forsyte.apalache.tla.lir.io.UTFPrinter
@@ -30,7 +31,7 @@ class Z3SolverContext(debug: Boolean = false, profile: Boolean = false) extends 
   private val z3context: Context = new Context()
   private val z3solver = z3context.mkSolver()
   private val simplifier = new ConstSimplifierForSmt()
-  private var smtListener: SmtListener = new FruitlessSmtListener()
+  private var smtListener: SmtListener = new IdleSmtListener()
 
   /**
     * Caching one uninterpreted sort for each cell signature. For integers, the integer sort.
